force bts and trusty percents + some refactor manage
